changing id type to long in both repositories

+++ OneThingApp/src/main/java/org/launchcode/OneThingApp/data/EntryRepository.java
@@ -3,6 +3,6 @@ package org.launchcode.OneThingApp.data;
 import org.launchcode.OneThingApp.models.Entry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntryRepository extends JpaRepository<Entry, Integer>{
+public interface EntryRepository extends JpaRepository<Entry, Long>{
 
 }
